Pull common components up a level
Connects to #792

The RelatedLinkComponent and RelatedLinkRowComponent are used both by works and by collections, so it is misleading to have these components in the app/components/works/ directly. Pull them up a level for clarity.
