+++ app/components/collections/form_component.html.erb
@@ -32,7 +32,7 @@
   </section>
 
   <section>
-    <%= render Works::RelatedLinkComponent.new(form: form) %>
+    <%= render RelatedLinkComponent.new(form: form) %>
   </section>
 
   <section>
