+++ app/components/related_link_component.rb
@@ -1,0 +1,11 @@
+# typed: true
+# frozen_string_literal: true
+
+# Renders a widget for describing a list of related links.
+class RelatedLinkComponent < ApplicationComponent
+  def initialize(form:)
+    @form = form
+  end
+
+  attr_reader :form
+end
