+++ app/components/related_link_row_component.html.erb
