+++ app/components/related_link_row_component.rb
@@ -1,0 +1,11 @@
+# typed: true
+# frozen_string_literal: true
+
+# Renders a widget for describing a related link.
+class RelatedLinkRowComponent < ApplicationComponent
+  def initialize(form:)
+    @form = form
+  end
+
+  attr_reader :form
+end
