+++ app/components/works/description_component.html.erb
@@ -83,6 +83,6 @@
 
   <%= render Works::RelatedWorkComponent.new(form: form) %>
 
-  <%= render Works::RelatedLinkComponent.new(form: form) %>
+  <%= render RelatedLinkComponent.new(form: form) %>
 
 </section>
