--- app/components/works/related_link_component.rb
@@ -1,13 +1,0 @@
-# typed: true
-# frozen_string_literal: true
-
-module Works
-  # Renders a widget for describing a list of related links.
-  class RelatedLinkComponent < ApplicationComponent
-    def initialize(form:)
-      @form = form
-    end
-
-    attr_reader :form
-  end
-end
